### 2025-09-19 #9024 @editorialbot
[PRE REVIEW]: floatCSEP: An application to deploy and conduct reproducible and prospective earthquake forecasting
**Submitting author:** @pabloitu (<a href="http://orcid.org/0000-0002-4787-1343">Pablo Cristián Iturrieta</a>)
**Repository:** https://github.com/cseptesting/floatcsep
**Branch with paper.md** (empty if default branch): main
**Version:** v0.3.0
