### 2025-10-14 #9024 @kthyng
@pabloitu ok, but please try not to add content in the review process. The paper is meant to stay high level and extra material can go in the docs.
