### 2025-11-07 #9024 @andreww
:wave: @scallag, @sscivier, @OlafZielke-EQ, @kirstybayliss, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html. Please be so kind to let me know even if you're not interested, so that I can start looking for other reviewers.

(Also, sorry to those of you I emailed ahead of finding your github usernames, perfectly happy for a reply here rather than by email if that's easiest)
