### 2025-11-11 757059e
build: placed build instructions into pyproject.toml. Removed setup.cfg and tox.ini
