+++ setup.cfg
@@ -1,91 +1,0 @@
-[metadata]
-name = floatcsep
-description = CSEP Floating Experiment application
-long_description = file: README.md
-long_description_content_type = text/markdown
-author = Pablo Iturrieta
-author_email = pciturri@gfz-potsdam.de
-license = BSD 3-Clause License
-version = attr: setuptools_scm.get_version
-platforms = unix, linux, osx, win32
-classifiers =
-    Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-url = https://github.com/cseptesting/floatcsep.git
-
-[options]
-packages =
-    floatcsep
-    floatcsep.commands
-    floatcsep.infrastructure
-    floatcsep.postprocess
-    floatcsep.utils
-install_requires =
-    numpy
-    dateparser
-    docker
-    gitpython
-    h5py
-    matplotlib
-    pandas
-    pycsep
-    pyshp
-    pyyaml
-    requests
-    scipy
-    seaborn
-    tables
-    xmltodict
-python_requires = >=3.9,<3.12
-zip_safe = no
-
-[options.packages.find]
-include = floatcsep*
-
-[options.extras_require]
-dev =
-    numpy
-    cartopy
-    black
-    dateparser
-    docker
-    flake8
-    geos
-    gitpython
-    h5py
-    matplotlib
-    mercantile
-    obspy
-    packaging
-    pandas
-    pycsep
-    pyproj
-    pyshp
-    pytest
-    pytest-cov
-    pyyaml
-    requests
-    scipy
-    seaborn
-    setuptools-scm
-    shapely
-    sphinx
-    sphinx-autoapi
-    sphinx_design
-    sphinx-toolbox
-    sphinx-gallery
-    sphinx-rtd-theme
-    sphinx_copybutton
-    tables
-    tox
-    vcrpy
-    xmltodict
-
-[options.entry_points]
-console_scripts =
-    floatcsep = floatcsep.commands.main:floatcsep
-
-[flake8]
-max-line-length = 96
