### 2025-11-11 8cf35e9
build: fallback to external tox.ini file. Removed setup.cfg from manifest. Modified LICENSE placement in .toml file. Added dev reqs to .toml. Updated requirement*.txt files. fix: released pymock requirements, which crashed numpy import.
