+++ tutorials/case_g/pymock/requirements.txt
@@ -1,3 +1,3 @@
-numpy==1.22.4
-matplotlib==3.4.3
-pytest==7.1.3
+numpy
+matplotlib
+pytest
