+++ MANIFEST.in
@@ -4,7 +4,6 @@ include requirements.txt
 include requirements_dev.txt
 include environment.yml
 include setup.py
-include setup.cfg
 include pyproject.toml
 include LICENSE
 include CREDITS.md
