+++ tox.ini
@@ -1,0 +1,26 @@
+[tox]
+minversion = 3.11
+envlist =
+    {py39,py310,py311}
+isolated_build = true
+requires = tox-conda
+
+[testenv]
+setenv =
+    PYTHONPATH = {toxinidir}
+conda_channels =
+    defaults
+    conda-forge
+conda_deps =
+    cartopy
+    pycsep
+deps =
+    -r{toxinidir}/requirements_dev.txt
+commands =
+    pip install -e .[dev]
+    pytest --basetemp={envtmpdir}
+package = wheel
+wheel_build_env = .pkg
+
+
+
