### 2025-11-13 #9024 @andreww
> Dear [@andreww](https://github.com/andreww). Thank you and the reviewers for being part of our submission. I have a small question: in this pre-review, the software version is pinned to 0.3.0, whereas there has been some development in between (to v0.4.1), mainly for visualizations, parallelization and new tutorials, albeit no API changes or new scientific features. Will the reviewers be able to review the latest version or it will be fixed to the version when the manuscript was submitted? Thank you!

Hi @pabloitu - I'm just about to kick off the review but I'll answer this first. The reviewers will generally look at the whole codebase but this mostly comes down to the latest version - so a combination of the tip of the main branch on github, whatever   pip (or conda) install installs, and whatever your latest documentation is. If we accept the submission one of the post-acceptance steps would be for you to create an up to date snapshot on something like zenodo for the accepted paper to point to. If there have been any code or documentation changes as part of the review this can sometimes be as simple as tagging a new point release on github and watching pypi and zenodo catch up.
