### 2025-11-13 #9408 @editorialbot
[REVIEW]: floatCSEP: An application to deploy and conduct reproducible, prospective earthquake forecasting experiments
**Submitting author:** @pabloitu (<a href="http://orcid.org/0000-0002-4787-1343">Pablo Cristián Iturrieta</a>)
**Repository:** https://github.com/cseptesting/floatcsep
**Branch with paper.md** (empty if default branch): floatcsep_joss
**Version:** v0.5.1
