### 2025-11-26 cseptesting/floatcsep#81 @sscivier
Contributions clarification
Issue raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist is

> **Contribution and authorship**: Has the submitting author (@pabloitu) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

For the first question, I am happy to answer yes. However, there appear to be some inconsistencies between the authors of the paper and the contributors to the `floatCSEP` package. Based on the information I could find, it was also not clear to me what are the contributions of each author. There is a `CREDITS.md` document in the repository--which is a good start, but it does not appear complete.

Specifics I have noticed:
- One contributor to `floatCSEP` code does not appear as an author on the paper.
- A number of authors on the paper do not appear as contributors to the code (in commit history); though non-code contributions can be sufficient for co-authorship. Some of these authors appear in `CREDITS.md`, but their contributions are not described.
- A number of authors on the paper do not appear in `CREDITS.md`.

In case I have missed something--is there somewhere in the paper / code / documentation where contributions and authorship are described?

If not, I would suggest the following:
- Ensure consistency between `CREDITS.md`, contributors to `floatCSEP`, and paper authors.
- Somewhere in the paper / code / documentation (likely `CREDITS.md`), briefly describe the contributions of all paper authors to the paper and/or the code.
- Following JOSS authorship guidelines (https://joss.readthedocs.io/en/latest/submitting.html#authorship), ensure that contributions are appropriately acknowledged / described--whether through co-authorship or acknowledgements in the paper.

These recommendations aim to help the authors meet JOSS authorship guidance; I’m happy for the authors to address them in whichever manner they consider best.
