### 2025-11-26 cseptesting/floatcsep#82 @sscivier
Paper summary
Issue raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist for the paper is:

> **Summary**: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

Overall, the summary communicates the software's high-level purpose, and is well-written and generally appropriate for the informed reader. However, I would suggest some minor adjustments to make it more digestible for a non-specialist reader.

- The terms "prospective" and "pseudo-prospective", in the context of earthquake forecasting, may be unfamiliar to a non-specialist (they were unfamiliar to me!). The authors may wish to re-phrase, or remove these terms.
- Would the authors consider adding a link to the `pyCSEP` toolkit in the summary (as is the case for the CSEP website)?
- (Very minor) The authors may consider slightly re-phrasing "...floatCSEP will support new official CSEP experiments..." to something like "...floatCSEP supports official CSEP experiments..." or "...floatCSEP is intended to support new official CSEP experiments..." to either highlight present benefit or qualify future promise, respectively.

I should note these are very minor comments, and the summary is well-written. I’m happy for the authors to address the above comments in whichever manner they consider best.
