### 2025-11-26 cseptesting/floatcsep#83 @sscivier
Paper: Statement of need
Issue raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist for the paper is:

> **A statement of need**: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

The **Statement of Need** is present, well-written, and mostly complete. It clearly communicates the problems addressed by the software, highlights gaps in existing software, and helps position `floatCSEP` relative to existing efforts.

However, my impression is that the target audience is more implied than specified. I believe the section would benefit from a brief, explicit statement of target audiences and users.

Additionally, the final sentence of the statement ("The solution should manage...") may benefit from specifying that `floatCSEP` is designed to address these requirements, and briefly highlighting how it overcomes previous issues. This would help explicitly situate `floatCSEP` among existing software.

These are minor comments, and I'm happy for the authors to address them in whichever manner they consider best.

EDIT:

I additionally noticed a double citation (`pandoc` syntax and manual text side-by-side):

> "...@mizrahi2024developing Mizrahi et al. (2024),..."
