### 2025-11-26 cseptesting/floatcsep#84 @sscivier
Paper: State of the field
Issue raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist for the paper is:

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

The paper does a good job of comparing `floatCSEP` to closely related projects (e.g., `pyCSEP`), and clearly explains how it fits into the existing CSEP software infrastructure. However, I think it would benefit from broader comparisons to other existing tools (outside CSEP) in this area, and perhaps wider seismology.

Suggestions:
- If other software packages exist outside the CSEP landscape for the kind of task that `floatCSEP` is designed for, then the paper would benefit from mentioning them and highlighting how `floatCSEP` differs.
- Adding some context around how `floatCSEP` integrates with (if it does) / complements other common seismology tools may help users understand how `floatCSEP` would fit into and / or add to their workflows.

I am happy for the authors to clarify if I have misunderstood anything in this regard, and address the above comment in whichever manner they consider best.
