### 2025-11-28 cseptesting/floatcsep#85 @sscivier
Docs: Statement of Need
This issue is raised as part of the JOSS review: openjournals/joss-reviews#9408

One of the items on my review checklist for the code documentation is:

> **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

The documentation is very well formatted, clearly outlines the purpose of `floatCSEP`, and explains the concepts behind the work. In relation to the checklist item above, I believe the documentation would benefit from (unless I had missed these somewhere):

- Explicit discussion of the gaps in existing tools that `floatCSEP` is addressing (similar to what is in the Statement of Need in the paper). Perhaps a section somewhere about "_Why `floatCSEP`?_" would address this--as an idea.
- Explicit statement of the target audiences and users. While implicitly clear, I couldn't find this being explicitly highlighted. Adding a small "_Target Audience / User_" section with 1-2 sentences could address this.

As with most of my review comments on this submission, these are minor and I'm happy for the authors to address them in whichever manner they consider best.
