### 2025-11-28 cseptesting/floatcsep#86 @sscivier
Docs: Example Usage
This issue is raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist for the documentation is:

> **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

The suite of tutorials is excellent, with many examples covering different scenarios. The tutorial documentation (`.rst` files) is also excellent, and the instructions are clear and simple to follow. However, the tutorial code lacks documentation, with minimal comments and docstrings. If a user were to try the tutorials only from the code, they may struggle to understand the examples fully. I believe the tutorial code would benefit from proper inline documentation and/or pointers to the `.rst` documentation on the docs site. The authors may additionally wish to highlight the tutorial docs in the **Run an Experiment** section of the `REAME.md`.
