### 2025-11-28 cseptesting/floatcsep#87 @sscivier
Docs: Functionality Documentation
This issue is raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist for the documentation is:

> **Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

The API documentation is generally excellent and comprehensive, with good coverage of main classes and functions, and good configuration documentation. However, there are some classes that appear to lack docstrings. I would recommend:

- Adding class- and method-level docstrings to the classes in `floatcsep.utils.file_io`.
- Adding a class-level docstring to `floatcsep.infrastructure.registries.ModelFileRegistry`.
- Adding a method-level docstring to `floatcsep.infrastructure.repositories.CatalogRepository.filter_catalog`.
