### 2025-11-28 cseptesting/floatcsep#88 @sscivier
Test coverage
This issue is raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist is:

> **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?

There is comprehensive test coverage for the core functionality, a good CI/CD setup, and multiple test types (e.g., unit, integration). However, the `postprocess` module appears to lack some coverage. In particular, the `postprocess/panel` module has no tests. I would encourage the authors to add some basic tests to increase coverage in the `postprocess` module and/or document if certain `postprocess` components are excluded from automated testing and why.
