### 2025-11-28 cseptesting/floatcsep#89 @sscivier
macOS system dependencies
This issue is raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items on my review checklist for the functionality is:

> **Installation**: Does installation proceed as outlined in the documentation?

The installation instructions are clear, and installation of the latest stable release from PyPI worked fine (I tested this using `uv` on macOS).

The documentation says that it is recommended--but not obligatory--to use a `conda` environment for developer installation. (Although this is rather specific) It doesn't explain that `pip`-only installation will fail on macOS (possibly other systems?) without system dependencies for WeasyPrint. Consider either: (a) making the `conda` recommendation stronger with an explanation, or (b) documenting the system dependency requirements for pip-based development installation.

For the editable/development installation using `pip`-only, the following steps worked for me on macOS:

```bash
# Install system dependencies first (required for WeasyPrint)
brew install glib pango gdk-pixbuf libffi

# Set library path (needed for runtime)
# This command is only active for the current terminal session. To make it persistent, add it to ~/.zshrc.
export DYLD_LIBRARY_PATH="/opt/homebrew/lib:$DYLD_LIBRARY_PATH"

# Clone and install
git clone https://github.com/cseptesting/floatcsep
cd floatcsep
uv pip install -e ".[dev]"

# Test installation
uv run floatcsep --help
uv run python -c "import floatcsep; print(floatcsep.__version__)"

# Run tutorial
cd tutorials/case_a
uv run floatcsep run config.yml
```

Note that I also required quotation marks for the development install to ensure `zsh` compatibility (this is shown without quotes in the installation docs currently):
```bash
pip install -e ".[dev]"
```
