### 2025-11-28 cseptesting/floatcsep#90 @sscivier
Lock file
Issue raised as part of the JOSS review: openjournals/joss-reviews#9408.

One of the items I have on my review checklist is:

> **Reproducibility**: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

The code and results are almost entirely reproducible, with: dependencies specified via e.g., `pyproject.toml` and `environment.yml`, good testing infrastructure, detailed docs, the `floatcsep reproduce` command, and the easy command line usage. However, I would recommend adding a lockfile (e.g., `conda-lock`, `uv.lock`) to pin exact dependency versions (including hashes) used during development and testing. This would help avoid issues owing to e.g., version drift.
