### 2025-12-07 e3aab06
repo: added detailed contributions to CREDITS.md. Added JOSS manuscript coauthors.
