+++ CREDITS.md
@@ -1,10 +1,76 @@
-Developer and scientific team:
+# Credits
 
-* Pablo Iturrieta - Deutsches GeoForschungsZentrum (GFZ), Germany - pciturri@gfz-potsdam.de
-* William Savran - University of Nevada, USA
-* Asim Khawaja - GFZ, Germany
-* Toño Bayona - University of Bristol, United Kingdom
-* Danijel Schorlemmer -  GFZ, Germany
-* Fabio Silva - Southern California Earthquake Center (SCEC), USA
-* Phil Maechling - SCEC, USA
-* Max Werner - University of Bristol, United Kingdom
+This project is the result of a collaborative effort by the following developer and scientific
+team.  
+Contributions include software development, scientific design, manuscript writing, testing,
+documentation, and project coordination.
+
+We describe contributions using short free-text descriptions and, where appropriate, the CRediT
+contributor roles
+(https://credit.niso.org/).
+
+## Developer and scientific team
+
+- **Pablo Iturrieta** – GFZ Helmholtz Centre for Geosciences, Germany –
+  pciturri@gfz-potsdam.de  
+  *Contributions:* [Conceptualization; Software (lead); Documentation; Methodology; Project administration; Formal Analysis; MS Writing – original draft]
+
+- **William Savran** – University of Nevada, USA  
+  *Contributions:* [Conceptualization; Software; Methodology; Formal Analysis; Validation; MS Writing – review & editing]
+
+- **Marcus Herrmann** – University of Naples, Italy  
+  *Contributions:* [Conceptualization; Software; Methodology; Documentation; Validation; MS Writing – review & editing]
+
+- **Toño Bayona** – University of Bristol, United Kingdom  
+  *Contributions:* [Methodology; Formal analysis; Validation; MS Writing – review & editing]
+
+- **Matt Gerstenberger** – Earth Sciences New Zealand  
+  *Contributions:* [Conceptualization; MS Writing – review & editing; Supervision]
+
+- **Kenny Graham** – Earth Sciences New Zealand  
+  *Contributions:* [Software; Methodology]
+
+- **Phil Maechling** – Southern California Earthquake Center (SCEC), USA
+  *Contributions:* [Conceptualization; Infrastructure; MS Writing – review & editing; Community coordination]
+
+- **Warner Marzocchi** – University of Naples, Italy  
+  *Contributions:* [Conceptualization; Supervision; MS Writing – review & editing]
+
+- **Leila Mizrahi** – ETH Zürich, Switzerland  
+  *Contributions:* [Software; Methodology]
+
+- **Danijel Schorlemmer** – GFZ, Germany  
+  *Contributions:* [Conceptualization; Supervision; Funding acquisition]
+
+- **Francesco Serafini** – University of Bristol, United Kingdom v  
+  *Contributions:* [Conceptualization; Methodology; Validation; MS Writing - review & editing]
+
+- **Fabio Silva** – SCEC, USA  
+  *Contributions:* [Validation; Infrastructure; Integration with SCEC systems]
+
+- **Asim Khawaja** – GFZ, Germany  
+  *Contributions:* [Software; Methodology]
+
+- **Max Werner** – University of Bristol, United Kingdom  
+  *Contributions:* [Conceptualization; Methodology; Formal Analysis; Supervision; MS Writing – review & editing; Community coordination; Funding Acquisition]
+
+---
+
+## Contribution roles (CRediT legend)
+
+- **Conceptualization** – Ideas; formulation or evolution of overarching research goals and
+  aims.
+- **Methodology** – Development or design of methods; creation of models or workflows.
+- **Software** – Programming, software development, implementation of code, or maintenance.
+- **Validation** – Verification of results, including code verification and benchmarking.
+- **Formal analysis** – Application of statistical, mathematical, or computational techniques.
[…]
