### 2025-12-07 4382e41
paper: remove prospective and pseudo-prospective tems from summary. Added pycsep link. Clarified floatCSEP is intended to be used for official CSEP experiments.
