### 2025-12-07 b03600f
paper: modified one double reference, and added target community.
