+++ paper.bib
@@ -6,8 +6,8 @@
   number={11},
   pages={2561--2562},
   year={2018},
-  doi={10.1073/pnas.180232411},
-  url={https://doi.org/10.1073/pnas.180232411},
+  doi={10.1073/pnas.1802324115},
+  url={https://doi.org/10.1073/pnas.1802324115},
   publisher={National Academy of Sciences}
 }
 @article{bayona2021two,
