### 2025-12-09 bf3e5a0
ms: fixed typo in summary and updated pdf
