### 2025-12-09 05fbf45
repo: added manuscript submission and its co-authorship to CREDITS.md
