+++ CREDITS.md
@@ -1,18 +1,14 @@
 # Credits
 
-This project is the result of a collaborative effort by the following developer and scientific
-team.  
-Contributions include software development, scientific design, manuscript writing, testing,
-documentation, and project coordination.
+This project is the result of a collaborative effort by the following developer and scientific team.
+Contributions include software development, scientific design, manuscript writing, testing, documentation, and project coordination.
+```floatCSEP``` is described in a manuscript submitted to Journal of Open-Science Software (https://joss.theoj.org/), whose authors were active collaborators at the time of submission.
 
-We describe contributions using short free-text descriptions and, where appropriate, the CRediT
-contributor roles
-(https://credit.niso.org/).
+We describe contributions using short free-text descriptions and, where appropriate, the CRediT contributor roles (https://credit.niso.org/).
 
 ## Developer and scientific team
 
-- **Pablo Iturrieta** – GFZ Helmholtz Centre for Geosciences, Germany –
-  pciturri@gfz-potsdam.de  
+- **Pablo Iturrieta** – GFZ Helmholtz Centre for Geosciences, Germany –   pciturri@gfz.de  
   *Contributions:* [Conceptualization; Software (lead); Documentation; Methodology; Project administration; Formal Analysis; MS Writing – original draft]
 
 - **William Savran** – University of Nevada, USA  
@@ -28,7 +24,7 @@ contributor roles
   *Contributions:* [Conceptualization; MS Writing – review & editing; Supervision]
 
 - **Kenny Graham** – Earth Sciences New Zealand  
-  *Contributions:* [Software; Methodology]
+  *Contributions:* [Software; Methodology; MS Writing – review & editing; ]
 
 - **Phil Maechling** – Southern California Earthquake Center (SCEC), USA
   *Contributions:* [Conceptualization; Infrastructure; MS Writing – review & editing; Community coordination]
@@ -37,7 +33,7 @@ contributor roles
   *Contributions:* [Conceptualization; Supervision; MS Writing – review & editing]
 
 - **Leila Mizrahi** – ETH Zürich, Switzerland  
-  *Contributions:* [Software; Methodology]
+  *Contributions:* [Software; Methodology; MS Writing – review & editing; ]
 
 - **Danijel Schorlemmer** – GFZ, Germany  
   *Contributions:* [Conceptualization; Supervision; Funding acquisition]
@@ -58,15 +54,13 @@ contributor roles
 
 ## Contribution roles (CRediT legend)
 
-- **Conceptualization** – Ideas; formulation or evolution of overarching research goals and
-  aims.
+- **Conceptualization** – Ideas; formulation or evolution of overarching research goals and aims.
 - **Methodology** – Development or design of methods; creation of models or workflows.
 - **Software** – Programming, software development, implementation of code, or maintenance.
 - **Validation** – Verification of results, including code verification and benchmarking.
 - **Formal analysis** – Application of statistical, mathematical, or computational techniques.
 - **Data curation** – Management of data, such as assembling catalogs or input datasets.
-- **Writing – original draft** – Preparation of the initial manuscript or main documentation
-  text.
+- **Writing – original draft** – Preparation of the initial manuscript or main documentation text.
 - **Writing – review & editing** – Critical review, commentary, and revision of text.
 - **Supervision** – Oversight and leadership responsibility, including mentoring.
 - **Project administration** – Coordination and management of the project.
