### 2025-12-09 ddd6151
ms: authors contribution linked to CREDITS.md
