### 2025-12-14 0d7a72b
docs: updated landing page
