### 2025-12-17 cseptesting/floatcsep#84 @sscivier
Thanks, @pabloitu. This makes sense. I agree that it's not worth detailing in-development integrations with other software. However, would it be possible to make the following aspects a bit more clear in the paper?
- To your knowledge, there is no software that supports the same use cases as `floatCSEP`.
- `floatCSEP` is designed to integrate with existing forecast-generating software.
- `floatCSEP` manages the entire end-to-end workflow, where other available software (including standard seismology tools / libraries) only cover parts of it.

Perhaps these could be highlighted with an extra 1-2 sentences, or by making minor modifications or additions to existing sentences in the section.

Please correct me if I’ve misunderstood any of the above aspects. I believe adding a bit more clarity to this section would help readers and users understand better where `floatCSEP` fits into the existing seismology software infrastructure. However, I don’t think this is essential if there simply isn’t room in the paper to add clarification.
