### 2025-12-22 709ec1c
docs: updated typo in landing page
