### 2025-12-22 d0ec14c
Updated MS with state of the field
