+++ paper.md
@@ -84,6 +84,7 @@ However, pyCSEP alone lacks key features required to deploy and conduct entire f
 That is, additional software is required that provides Testing Center capabilities while remaining decoupled from specific hosting hardware.
 To meet this need, we developed floatCSEP, which manages the entire experiment lifecycle, from model integration and initial deployment to the incremental updating of input data, forecasts, results, and reports as new observations become available.
 It is intended for earthquake forecast model developers, institutions that run CSEP-style forecasting experiments, and the broader statistical seismology community.
+To our knowledge, no existing software provides this complete end-to-end testing workflow; commonly used seismology tools instead address only individual steps, such as catalog queries (e.g., [ObsPy](https://github.com/obspy/obspy))) or forecast evaluation (e.g., pyCSEP).
 
 # Software Overview
 
@@ -96,6 +97,7 @@ This declarative approach simplifies the experiment setup, standardizes its work
 floatCSEP uses pyCSEP as a dependency, incorporating its core functionality (forecast and catalog classes, and evaluation routines) alongside additional Testing Center operations, such as data management and computational containerization.
 The application supports multiple forecast formats and accommodates both time-invariant and time-variant experiments.
 It handles forecasts produced either by models managed directly by floatCSEP or provided externally through raw files.
+Rather than replacing or modifying existing forecast-generating codes, floatCSEP is designed to integrate with them and run them as part of a unified, automated workflow.
 Representative use cases are included as tutorials, which users can extend by incorporating new models, adding alternative evaluations, or by replicating in different regions and timeframes.
 The software integrates seamlessly with pyCSEP’s existing testing routines, but also provides custom hooks for user-defined tests, visualizations, and reports.
 
