### 2025-12-22 dbd092c
tutorials: removed python scripts to run tutorials. Added tutorial documentation link to README quick usage.
