### 2025-12-23 eb28cd4
docs: Added docstrings to file_io module
