+++ floatcsep/utils/file_io.py
@@ -4,46 +4,114 @@ import logging
 import os.path
 import time
 import xml.etree.ElementTree as eTree
+from typing import Iterator
 
 import csep
 import h5py
 import numpy
 import pandas
-import pandas as pd
 from csep.core.catalogs import CSEPCatalog
 from csep.core.regions import QuadtreeGrid2D, CartesianGrid2D
+from csep.core.forecasts import CatalogForecast
 from csep.models import Polygon
 from csep.utils.time_utils import strptime_to_utc_epoch
+
 
 log = logging.getLogger(__name__)
 
 
 class CatalogSerializer:
+    """
+    Serializers for :class:`csep.core.catalogs.CSEPCatalog`.
+
+    Delegates to the built-in I/O methods from pyCSEP catalog objects.
+    """
 
     @staticmethod
-    def ascii(catalog, filename: str) -> None:
+    def ascii(catalog: CSEPCatalog, filename: str) -> None:
+        """
+        Serialize a catalog to the pyCSEP ASCII format.
+
+        Args:
+            catalog (:class:`csep.core.catalogs.CSEPCatalog`): Catalog instance to write.
+            filename (str): Destination filepath.
+
+        Returns:
+            None
+        """
         catalog.write_ascii(filename=filename)
 
     @staticmethod
-    def json(catalog, filename: str) -> None:
+    def json(catalog: CSEPCatalog, filename: str) -> None:
+        """
+        Serialize a catalog to JSON using the pyCSEP encoder.
+
+        Args:
+            catalog (:class:`csep.core.catalogs.CSEPCatalog`): Catalog instance to write.
+            filename (str): Destination filepath.
+
+        Returns:
+            None
+        """
         catalog.write_json(filename=filename)
 
 
 class CatalogParser:
+    """
+    Parsers for :class:`csep.core.catalogs.CSEPCatalog`.
 
+    Wraps pyCSEP catalog loaders and provides an interface for loading catalogs.
+    """
     @staticmethod
-    def ascii(filename: str) -> None:
+    def ascii(filename: str) -> CSEPCatalog:
+        """
+        Load a pyCSEP catalog in the ASCII catalog format.
+
+        Args:
+            filename (str): Path to file.
+
+        Returns:
+            :class:`csep.core.catalogs.CSEPCatalog`.
+        """
         return CSEPCatalog.load_catalog(filename=filename)
 
     @staticmethod
-    def json(filename: str) -> None:
+    def json(filename: str) -> CSEPCatalog:
+        """
+        Load a pyCSEP catalog from JSON.
+
+        Args:
+            filename (str): Path to the file.
+
+        Returns:
+            :class:`csep.core.catalogs.CSEPCatalog`
+        """
         return CSEPCatalog.load_json(filename=filename)
 
 
 class CatalogForecastParsers:
+    """
+    Parsers for catalog-based forecasts stored on disk.
 
+    These helpers select the appropriate loader based on file headers and return
+    the object produced by :func:`csep.load_catalog_forecast`.
+    """
     @staticmethod
-    def csv(filename, **kwargs):
+    def csv(filename: str, **kwargs) -> CatalogForecast:
+        """
+        Load a catalog-based forecast from a CSV file.
+
+        The loader inspects the header to detect either:
+        (i) a CSEP/pyCSEP-style catalog layout or (ii) a Hermes layout.
+
+        Args:
+            filename (str): Path to the CSV file.
+            **kwargs: Passed to :func:`csep.load_catalog_forecast`.
+
+        Returns:
+            :class:`csep.core.forecasts.CatalogForecast`.
+
+        """
         csep_headers = [
             "lon",
             "lat",
@@ -61,7 +129,7 @@ class CatalogForecastParsers:
             "longitude",
             "time",
         ]
[…]
