### 2025-12-23 16724f9
docs: added class docstrings to ModelFileRegistry.
