### 2025-12-23 bc7be11
docs: added method docstring to CatalogRepository.filter_catalog
