### 2025-12-27 afad123
docs: updated instructions for pip-only instructions, including the required binary libraries for debian/ubuntu and macos. ci: added smoke builds for pip-only installations
