+++ .github/workflows/build-test.yml
@@ -8,7 +8,7 @@ on:
     branches: [ main ]
 
 jobs:
-  build:
+  build-test:
     if: github.repository == 'cseptesting/floatcsep'
     runs-on: ${{ matrix.os }}
     strategy:
@@ -54,3 +54,61 @@ jobs:
         with:
           token: ${{ secrets.CODECOV_TOKEN }}
           fail_ci_if_error: false
+
+  pip_smoke_linux:
+    if: github.repository == 'cseptesting/floatcsep'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4.2.2
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+          cache: pip
+
+      - name: Install system deps (WeasyPrint)
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y \
+            libglib2.0-0 \
+            libpango-1.0-0 \
+            libpangoft2-1.0-0 \
+            libharfbuzz0b \
+            libharfbuzz-subset0
+
+      - name: Pip-only install (no tests)
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install -e ".[dev]"
+          python -c "import floatcsep; print('floatcsep OK:', floatcsep.__version__)"
+          python -c "import weasyprint; print('WeasyPrint OK:', weasyprint.__version__)"
+          floatcsep --help
+
+
+  pip_smoke_macos:
+    if: github.repository == 'cseptesting/floatcsep'
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@v4.2.2
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+          cache: pip
+
+      - name: Install system deps (WeasyPrint)
+        run: |
+          brew update
+          brew install cairo pango gdk-pixbuf libffi glib
+
+      - name: Pip-only install (no tests)
+        env:
+          DYLD_FALLBACK_LIBRARY_PATH: /opt/homebrew/lib:/usr/local/lib:${{ env.DYLD_FALLBACK_LIBRARY_PATH }}
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install -e ".[dev]"
+          python -c "import floatcsep; print('floatcsep OK:', floatcsep.__version__)"
+          python -c "import weasyprint; print('WeasyPrint OK:', weasyprint.__version__)"
+          floatcsep --help
