### 2025-12-27 4c07435
ci: clean mamba cache for tests
