### 2025-12-27 af06b77
build: added conda lockfile to check dependencies pertaining to a given release. ci: test floatcsep build with the given release lockfile.
