### 2025-12-27 cseptesting/floatcsep#103 @pabloitu
JOSS review - References
This issue tracks a comment by @sscivier from the JOSS review thread.

Source (JOSS): https://github.com/openjournals/joss-reviews/issues/9408#issuecomment-3663168998

> Regarding the **References** checklist item for the paper, I'm nearly happy to check it. A couple of minor things:
>
>     * As mentioned in [Paper: Statement of need cseptesting/floatcsep#83](https://github.com/cseptesting/floatcsep/issues/83), there is a duplicate `pandoc` / manual text citation in the **Statement of Need**--just a typo and the manual text needs removing.
>
>     * The paper mentions containerised environments as a feature, but does not cite a particular dependency/spec. If the paper relies on a specific dependency (e.g., `Docker`) for containerisation, the authors may wish to cite this.
