### 2025-12-27 cseptesting/floatcsep#105 @pabloitu
JOSS review - bug tracker
This issue tracks a comment by @sscivier from the JOSS review thread.

Source (JOSS): https://github.com/openjournals/joss-reviews/issues/9408#issuecomment-3582017126

> Minor note on `CONTRIBUTING.md`--the link for the "bug tracker" didn't work for me. Otherwise, I'm happy to tick the **Community guidelines** item for the software.
