### 2025-12-27 c50877f
repo: fixed typo in link to the bug tracker in CONTRIBUTING.md
