### 2025-12-27 cseptesting/floatcsep#107 @pabloitu
JOSS review - Figure 2 contrast
This issue tracks a comment by @sscivier from the JOSS review thread.

Source (JOSS): https://github.com/openjournals/joss-reviews/issues/9408#issuecomment-3585279801

> A quick note on Fig 2 of the paper. I think it works well to have an example `config.yml` file. However, there is not sufficient contrast between the text and background colours, which makes it difficult to read.
