### 2025-12-28 cseptesting/floatcsep#107 @pabloitu
Updated figure 2 contrast with different colors for the background and text to improve visibility. Changes added in #108
