### 2025-12-28 cseptesting/floatcsep#109 @pabloitu
JOSS review - CITATION.cff
This issue tracks a comment by @sscivier from the JOSS review thread.

Source (JOSS): https://github.com/openjournals/joss-reviews/issues/9408#issuecomment-3589654040

> I would encourage the authors to add a `CITATION.cff` file to the repo. This would:
>
>     * Make it easy for users to cite their use of the software, with a "_Cite this repository_" button automatically added to the repository side bar on GitHub.
>
>     * Formalise and make the metadata machine-readable.
>
>     * Allow users to cite the JOSS paper and the software itself with its Zenodo DOI. Within the `CITATION.cff` file, the authors may specify a `preferred-citation` that could point to the JOSS paper.
>
>     * Allow for e.g., ORCiDs to be properly linked, and ensure all authors are credited appropriately.
>
>     * Enable version specific citations consistent with Zenodo.
>
>
> In line with this, I would encourage the authors to add a "**How to cite**" (or similar) section to the `README.md`.
