### 2025-12-28 7df4a11
repo: added CITATION.cff for machine readable search. Added small section "How to CIte" to README.md
