### 2025-12-28 cseptesting/floatcsep#109 @pabloitu
Thank you for this info! I added a CITATION.cff and included the into in the README.md,  now linking to the zenodo repo, but hopefully to the JOSS paper once its published. Changes added in #110
