### 2025-12-30 cseptesting/floatcsep#86 @sscivier
Thanks, @pabloitu. I'll be happy to tick off this item on my checklist once the PR #98 is merged.
