### 2025-12-30 cseptesting/floatcsep#87 @sscivier
Thanks, @pabloitu. I'll be happy to tick this item off my checklist once the PR #99 is merged.
