### 2025-12-30 #9408 @sscivier
@pabloitu -- Thanks for comprehensively addressing the review comments. I believe the changes have strengthened the submission. As a summary, I have put below the (currently) remaining checklist items, and the actions needed for me to tick them off.

_General Checks_

- [ ] **Contribution and Authorship**: Merge PRs cseptesting/floatcsep#93 and cseptesting/floatcsep#94.
- [ ] **Reproducibility**: Address the comment: https://github.com/cseptesting/floatcsep/issues/109#issuecomment-3698360736, **then** merge PRs cseptesting/floatcsep#102 and cseptesting/floatcsep#110 **after** making any updates as necessary.

_Functionality_

- [ ] **Installation**: Merge PR cseptesting/floatcsep#101.

_Documentation_

- [ ] **Statement of Need**: Merge PR cseptesting/floatcsep#96.
- [ ] **Example Usage**: Merge PR cseptesting/floatcsep#98.
- [ ] **Functionality Documentation**: Merge PR cseptesting/floatcsep#99.
- [ ] **Automated Tests**: Merge PR cseptesting/floatcsep#100.
- [ ] **Community Guidelines**: Merge PR cseptesting/floatcsep#106.

_Paper_

- [ ] **Summary**: Merge PR cseptesting/floatcsep#92.
- [ ] **Statement of Need**: Merge PR cseptesting/floatcsep#95.
- [ ] **State of the Field**: Merge PR cseptesting/floatcsep#97.
- [ ] **References**: Merge PR cseptesting/floatcsep#104.

_Other_

- [ ] Merge PR cseptesting/floatcsep#108.
