### 2025-12-30 cseptesting/floatcsep#109 @andreww
My view is that it's up to the authors to point users to what they think is the best citation for their software. Having said that, having something to cite is one of the reasons JOSS exists.

I would certainly make sure people looking to cite floatCSEP get a consistent message from the README and cff files. Part of the process after review is for me to check all the versions and links tie together properly but JOSS doesn't mint a doi you can cite until we actually publish the paper. The thing I think most people do is put something reasonable (e.g. a zenodo archive) in the citation links and then, if they want the JOSS paper to be the thing people should cite, update the github repository to point to that once the JOSS paper is out.
