### 2025-12-30 ade96a4
tests: skip case_d on CI, due to network issues with Zenodo with CI workflow for multiple OS/python versions.
