### 2025-12-30 3491f6b
Modified manuscript's summary (#92)
* paper: remove prospective and pseudo-prospective tems from summary. Added pycsep link. Clarified floatCSEP is intended to be used for official CSEP experiments.

* paper: modified one double reference, and added target community.

* ms: fixed typo in summary and updated pdf
