### 2025-12-30 0ad5c28
Updated MS Statement of Need #95
ms: explained how floatcsep fits the need and  which is its target audience (#95)
