### 2025-12-30 c6aa646
Updated JOSS manuscript with state of the field / models' code integration (#97)
* Updated MS with state of the field

* ms: fixed small parenthesis typo
