### 2025-12-30 5ffa36c
ms: added Docker reference (#104)
