+++ paper.bib
@@ -190,6 +190,18 @@
   url={https://doi.org/10.1785/0120220028},
   publisher={Seismological Society of America}
 }
+@article{merkel2014docker,
+author = {Merkel, Dirk},
+title = {Docker: lightweight Linux containers for consistent development and deployment},
+year = {2014},
+issue_date = {March 2014},
+publisher = {Belltown Media},
+address = {Houston, TX},
+volume = {2014},
+number = {239},
+issn = {1075-3583},
+journal = {Linux Journal},
+}
 @article{michael2018preface,
   title={Preface to the focus section on the Collaboratory for the Study of Earthquake Predictability (CSEP): New results and future directions},
   author={Michael, Andrew J and Werner, Maximilian J},
