### 2025-12-30 3a8529b
ms: updated figure 2 with better contrast between background and text (#108)
* ms: updated figure 2 with better contrast between background and text

* ms: updated ms pdf
