### 2025-12-30 30ba79a
repo: updated CITATION.cff and citation instructions in README.md
