+++ CITATION.cff
@@ -16,51 +16,18 @@ authors:
     given-names: "Pablo"
     orcid: "https://orcid.org/0000-0002-4787-1343"
     affiliation: "GFZ Helmholtz Centre for Geosciences, Potsdam, Germany"
+
+  - family-names: "Khawaja"
+    given-names: "Asim M."
+    orcid: "https://orcid.org/0000-0002-6196-4503"
+    affiliation: "GFZ Helmholtz Centre for Geosciences, Potsdam, Germany"
+
   - family-names: "Savran"
     given-names: "William H."
     orcid: "https://orcid.org/0000-0001-5404-0778"
     affiliation: "University of Nevada, Reno, United States"
-  - family-names: "Herrmann"
-    given-names: "Marcus"
-    orcid: "https://orcid.org/0000-0002-2342-1970"
-    affiliation: "Università degli Studi di Napoli Federico II, Naples, Italy"
-  - family-names: "Bayona"
-    given-names: "José A."
-    orcid: "https://orcid.org/0000-0002-3026-0327"
-    affiliation: "School of Earth Sciences, University of Bristol, Bristol, United Kingdom"
-  - family-names: "Gerstenberger"
-    given-names: "Matt C."
-    orcid: "https://orcid.org/0000-0002-2938-1359"
-    affiliation: "GNS Science | Te Pū Ao, Lower Hutt, New Zealand"
-  - family-names: "Graham"
-    given-names: "Kenny M."
-    orcid: "https://orcid.org/0000-0003-3645-6344"
-    affiliation: "GNS Science | Te Pū Ao, Lower Hutt, New Zealand"
-  - family-names: "Maechling"
-    given-names: "Philip J."
-    orcid: "https://orcid.org/0000-0002-9221-7068"
-    affiliation: "Southern California Earthquake Center, University of Southern California, United States"
-  - family-names: "Marzocchi"
-    given-names: "Warner"
-    orcid: "https://orcid.org/0000-0002-9114-1516"
-    affiliation: "Università degli Studi di Napoli Federico II, Naples, Italy"
-  - family-names: "Mizrahi"
-    given-names: "Leila"
-    orcid: "https://orcid.org/0000-0002-5262-3168"
-    affiliation: "Swiss Seismological Service at ETH Zürich, Zürich, Switzerland"
+
   - family-names: "Schorlemmer"
     given-names: "Danijel"
     orcid: "https://orcid.org/0000-0003-3969-1059"
     affiliation: "GFZ Helmholtz Centre for Geosciences, Potsdam, Germany; Swiss Seismological Service at ETH Zürich, Zürich, Switzerland"
-  - family-names: "Serafini"
-    given-names: "Francesco"
-    orcid: "https://orcid.org/0000-0003-0154-6200"
-    affiliation: "School of Earth Sciences, University of Bristol, Bristol, United Kingdom"
-  - family-names: "Silva"
-    given-names: "Fabio"
-    orcid: "https://orcid.org/0000-0002-6543-5705"
-    affiliation: "Southern California Earthquake Center, University of Southern California, United States"
-  - family-names: "Werner"
-    given-names: "Maximilian J."
-    orcid: "https://orcid.org/0000-0002-2430-2631"
-    affiliation: "School of Earth Sciences, University of Bristol, Bristol, United Kingdom"
