### 2025-12-30 cseptesting/floatcsep#109 @pabloitu
I've made the CITATION.cff consistent with the latest Zenodo release's metadata (only 4 authors present at the moment in which the software was created 3 years ago). I'm happy to update the CITATION.cff so it points its "Preferred citation" to the JOSS article once it can be published.

Additionally, I've updated the README.md instructing to use the "Cite this repository" button, but also placed a consistent citation text beneath.

* Note: If the Zenodo version doesn't make sense to you, it is because the zenodo-github webhook delivery broke after v0.3.0, and after some research and unfruitful debugging, it is likely due to not having a CITATION.cff in the repo. Thus I find myself in a small circular conundrum together with this raised issue, but I hope you agree with the solution proposed above.
