### 2026-01-02 fb7bee9
ms: increased figure 1 resolution. Fixed a wrong ORCID link. Removed accidental copy-paste from manuscript. Updated figure 2 caption. Updated ms pdf. Updated link to CREDITS.md from the main branch.
