### 2026-01-02 2171cea
repo: updated CITATION.cff with the fixed manuscript title
