### 2026-01-02 cseptesting/floatcsep#82 @pabloitu
Closed with #92
