### 2026-01-02 cseptesting/floatcsep#84 @pabloitu
Closed with #97
