### 2026-01-05 c05232f
ms: fixed manuscript title and added co-authors ORCID. (#114)
* ms: fixed manuscript title and added co-authors ORCID.

* ms: increased figure 1 resolution. Fixed a wrong ORCID link. Removed accidental copy-paste from manuscript. Updated figure 2 caption. Updated ms pdf. Updated link to CREDITS.md from the main branch.
