+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'floatCSEP: An application to deploy and conduct reproducible and prospective earthquake forecasting'
+title: 'floatCSEP: An application to deploy and conduct reproducible, prospective earthquake forecasting experiments'
 tags:
   - Python
   - seismology
@@ -11,28 +11,40 @@ authors:
     corresponding: true
     affiliation: 1
   - name: William H. Savran
+    orcid: 0000-0001-5404-0778
     affiliation: 2
   - name: Marcus Herrmann
+    orcid: 0000-0002-2342-1970
     affiliation: 3
   - name: José A. Bayona
+    orcid: 0000-0002-3026-0327
     affiliation: 4
-  - name: Matt C. Gerstenberger
+  - name: Matthew C. Gerstenberger
+    orcid: 0000-0002-0392-7114
     affiliation: 5
-  - name: Kenny M. Graham
+  - name: Kenny Graham
+    orcid: 0000-0003-3645-6344
     affiliation: 5
   - name: Philip J. Maechling
+    orcid: 0000-0002-9221-7068
     affiliation: 6
   - name: Warner Marzocchi
+    orcid: 0000-0002-9114-1516
     affiliation: 3
   - name: Leila Mizrahi
+    orcid: 0000-0002-5262-3168
     affiliation: 7
   - name: Danijel Schorlemmer
+    orcid: 0000-0003-3969-1059
     affiliation: "1, 7"
   - name: Francesco Serafini
+    orcid: 0000-0003-0154-6200
     affiliation: 4
   - name: Fabio Silva
+    orcid: 0000-0002-6543-5705
     affiliation: 6
   - name: Maximilian J. Werner
+    orcid: 0000-0002-2430-2631
     affiliation: 4
 affiliations:
   - name: GFZ Helmholtz Centre for Geosciences, Potsdam, Germany,
@@ -105,7 +117,7 @@ The software integrates seamlessly with pyCSEP’s existing testing routines, bu
 
 An example configuration file (`config.yml`) for a time-invariant, grid-based experiment in Italy with two models is shown in \autoref{fig:example}.
 
-![Simplest example of a configuration file for a time-invariant, grid-based experiment with two models. The `run` command executes an experiment end-to-end. The `stage` command accesses and builds the models' source code and prepares input/testing catalogs. The `reproduce` command re-executes an experiment and compares it with existing results using statistical and computational metrics. `plot` executes the post-process and visualization of the experiment; The `update` command generates and tests all forecasts missing since the last execution up to today. \label{fig:example}](figures/fig2_example_config.png){width="350pt"}
+![Simplest example of a configuration file for a time-invariant, grid-based experiment with two models. The `run` command executes the experiment end-to-end, generating and testing all missing forecasts from the last execution up to today. The `stage` command accesses and builds the models' source code, as well as preparing input and testing catalogs. The `reproduce` command re-executes an experiment and compares it with existing results using statistical and computational metrics. `plot` executes the post-process, image generation, and reporting of the experiment. `view` deploys a data app for visualizing the forecasts, catalogs, and test results. \label{fig:example}](figures/fig2_example_config.png){width="350pt"}
 
 # Applications
 
@@ -117,11 +129,11 @@ floatCSEP is designed to support the following applications:
 - Plug in new models into a completed or ongoing (float)CSEP experiment. Since CSEP experiments are clearly defined, they can be effectively used as benchmarks for comparing and developing new forecasting models [e.g., @serafini2025benchmark]
[…]
