### 2026-01-06 #9408 @sscivier
**Minor clarity issue**:

> _Line 74_: "...environment that can run on any computer with sufficient computational resources..."

The phrase “sufficient computational resources” is vague and could span anything from a laptop to an HPC system. Would the authors consider briefly specifying minimum or typical resource requirements (e.g., CPU cores, memory), or rephrasing with a more specific qualitative constraint (e.g., standard workstation).
