### 2026-01-06 #9408 @sscivier
@pabloitu -- Thanks for all your work in addressing the review comments. I have now ticked off all of the items on my checklist. I read through the current proof, and suggested some minor updates in my comments above. Once these comments have been addressed, I will be happy to recommend acceptance.
