### 2026-01-19 #9408 @pabloitu
@sscivier Thank you for the final reviews! I added your recommendations with commit cseptesting/floatcsep@696ffb6eae19c0ab943f82885207e5dac81f8dad

Regarding the review:

> **Minor clarity issue**:
>
> > _Line 74_: "...environment that can run on any computer with sufficient computational resources..."
>
> The phrase “sufficient computational resources” is vague and could span anything from a laptop to an HPC system. Would the authors consider briefly specifying minimum or typical resource requirements (e.g., CPU cores, memory), or rephrasing with a more specific qualitative constraint (e.g., standard workstation).

I would prefer not to hard-specify quantitative resources, since it is highly dependent on the models runtime optimization, the forecast window, and the experiment expected duration. Most short/research(retrospective)/time-unvarying experiments could be run in a personal workstation, but in official, complex and long-breath experiments it would be optimal to have a dedicated server. We modify the text to briefly clarify:

> The primary objective of floatCSEP is to provide a portable, automated, and reproducible Testing Center environment that runs on a standard workstation (subject to experiment-specific runtime and memory demands) and scales to HPC systems for computationally intensive experiments, such as time-dependent models producing daily forecasts.

Thank you again for such dedicated reviews.
