### 2026-01-23 ada0be8
fix: zenodo queries are now done per file per model (flavour). Reduced query rate to avoid being flagged by zenodo's api. Added throttle and captured more exceptions caused by other server responses. Avoid forceful download if file already existed. tutorials: reduced amount of models for zenodo's tutorial d).
