+++ floatcsep/utils/accessors.py
@@ -1,66 +1,100 @@
+import time
 import git
 import requests
 import hashlib
 import os
-import sys
 import shutil
 
 
-def from_zenodo(record_id, folder, force=False):
-    """
-    Download data from a Zenodo repository.
+def from_zenodo(record_id, folder, force=False, keys=None):
+    record_url = f"https://zenodo.org/api/records/{record_id}"
+    max_tries = 5
 
-    Downloads if file does not exist, checksum has changed in local respect to url or force
+    os.makedirs(folder, exist_ok=True)
 
-    Args:
-        record_id: corresponding to the Zenodo repository
-        folder: where the repository files will be downloaded
-        force: force download even if file exists and checksum passes
+    for attempt in range(1, max_tries + 1):
+        r = requests.get(record_url, timeout=30, headers={"User-Agent": "floatcsep"})
 
-    Returns:
-    """
-    # Grab the urls and filenames and checksums
-    r = requests.get(f"https://zenodo.org/api/records/{record_id}", timeout=30)
-    download_urls = [f["links"]["self"] for f in r.json()["files"]]
-    filenames = [(f["key"], f["checksum"]) for f in r.json()["files"]]
+        if r.status_code == 200:
+            break
 
-    # Download and verify checksums
+        if r.status_code == 403:
+            text = (r.text or "").lower()
+            if "unusual traffic" in text or "<html" in text:
+                snippet = (r.text or "")[:400].replace("\n", "\\n")
+                raise RuntimeError(
+                    "Zenodo returned HTTP 403 and appears to be blocking this network/IP due "
+                    "to unusual traffic.\n"
+                    f"URL: {record_url}\n"
+                    f"Response snippet: {snippet}"
+                )
+            r.raise_for_status()
+
+        if r.status_code in (429, 500, 502, 503, 504):
+            wait = min(2 ** (attempt - 1), 30)
+            ra = r.headers.get("Retry-After")
+            if ra:
+                try:
+                    wait = max(wait, int(ra))
+                except ValueError:
+                    pass
+            time.sleep(wait)
+            continue
+
+        r.raise_for_status()
+    else:
+        raise RuntimeError(f"Zenodo API request failed after {max_tries} attempts: {record_url}")
+
+    try:
+        data = r.json()
+    except Exception as e:
+        snippet = (r.text or "")[:400].replace("\n", "\\n")
+        raise RuntimeError(
+            "Zenodo API did not return valid JSON.\n"
+            f"URL: {record_url}\n"
+            f"Content-Type: {r.headers.get('Content-Type')!r}\n"
+            f"Snippet: {snippet}"
+        ) from e
+
+    files = data.get("files", [])
+    if not isinstance(files, list):
+        raise RuntimeError(f"Zenodo record JSON missing expected 'files' list: {record_url}")
+
+    if keys is not None:
+        wanted = set(keys)
+        files = [f for f in files if f.get("key") in wanted]
+        missing = wanted - {f.get("key") for f in files}
+        if missing:
+            raise FileNotFoundError(
+                f"Zenodo record {record_id} does not contain required file(s): {sorted(missing)}"
+            )
+
+    download_urls = [f["links"]["self"] for f in files]
+    filenames = [(f["key"], f["checksum"]) for f in files]
+
     for (fname, checksum), url in zip(filenames, download_urls):
         full_path = os.path.join(folder, fname)
+
         if os.path.exists(full_path):
             value, digest = check_hash(full_path, checksum)
             if value != digest:
-                print(f"Checksum is different: re-downloading {fname}" f" from Zenodo...")
+                print(f"Checksum differs, re-downloading {fname} ...")
                 download_file(url, full_path)
             elif force:
-                print(f"Re-downloading {fname} from Zenodo...")
[…]
