### 2026-01-23 811e7c5
fix: removed run.py objects from the Dockerfiles in case_i, which were removed from the model repository. fix: timedependent get_source() are able to clone into empty folder.
