### 2026-01-23 fc53174
117 zenodo call throws error due to request rate limit (#118)
* fix: zenodo queries are now done per file per model (flavour). Reduced query rate to avoid being flagged by zenodo's api. Added throttle and captured more exceptions caused by other server responses. Avoid forceful download if file already existed.
tutorials: reduced amount of models for zenodo's tutorial d).

* fix: removed run.py objects from the Dockerfiles in case_i, which were removed from the model repository.
fix: timedependent get_source() are able to clone into empty folder.
