+++ tutorials/case_d/models.yml
@@ -5,11 +5,6 @@
       N10L11: TEAM=N10L11.csv
       N25L11: TEAM=N25L11.csv
       N50L11: TEAM=N50L11.csv
-      N100L11: TEAM=N100L11.csv
-      SN10L11: TEAM=SN10L11.csv
-      SN25L11: TEAM=SN25L11.csv
-      SN50L11: TEAM=SN50L11.csv
-      SN100L11: TEAM=SN100L11.csv
 - WHEEL:
     zenodo_id: 6255575
     path: models/wheel
@@ -17,8 +12,3 @@
       N10L11: WHEEL=N10L11.csv
       N25L11: WHEEL=N25L11.csv
       N50L11: WHEEL=N50L11.csv
-      N100L11: WHEEL=N100L11.csv
-      SN10L11: WHEEL=SN10L11.csv
-      SN25L11: WHEEL=SN25L11.csv
-      SN50L11: WHEEL=SN50L11.csv
-      SN100L11: WHEEL=SN100L11.csv
