### 2026-01-23 590f86f
Update MANIFEST.in (#119)
