+++ MANIFEST.in
@@ -1,5 +1,6 @@
 include README.md
 include CODE_OF_CONDUCT.md
+include CITATION.cff
 include requirements.txt
 include requirements_dev.txt
 include environment.yml
