### 2026-01-27 #9408 @andreww
> [@pabloitu](https://github.com/pabloitu) -- Thanks for those final edits.
>
> [@andreww](https://github.com/andreww) -- I'm happy to recommend acceptance.

Thanks for reviewing @sscivier!
