### 2026-01-28 cseptesting/floatcsep#120 @pabloitu
Improve documentation related to pyCSEP
From the [JOSS review](https://github.com/openjournals/joss-reviews/issues/9408#issuecomment-3716884474) by @keliankaz

> One concern is that it may be difficult for those not very familiar with the pyCSEP ecosystem to get up an running with a floating experiment. This is unavoidable as floatCSEP is built around pyCSEP. Still, I wonder if providing a few more details in the tutorials to clearly explain aspects of the pyCSEP api that are being used would alleviate this issue. In most cases the required documentation is available in wither the "Defining an experiment" section or in the pyCSEP documentation.

The ```floatcsep``` documentation should provide a better overview of how each tutorial is interfacing with ```pycsep`` (e.g., which test function will the tutorial using, which forecast classes are involved, which catalog query functions or region readers we are using, etc.). Simultaneously, more clear links to the function/classes involved should be placed within each tutorial.
