### 2026-01-29 f61b578
docs: updated every tutorial with additional links to pycsep functions/classes. Added hints pointing users to advance documentation when a new concept is introduced. Added an extra section `pyCSEP under the hood`, so users can see which pyCSEP functions where used in the tutorials. Upgraded some typos in cases g and h, due to config.yml updates. Removed depreacted `use_db` from case_e.rst.
