+++ docs/tutorials/case_h.rst
@@ -76,12 +76,12 @@ As in :ref:`Tutorial G<case_g>`, each **Model** requires to build and execute a 
 .. note::
     The ``models.yml`` will define how to interface **floatCSEP** to each Model, implying that a Model should be developed, or adapted to ensure the interface requirements specified below.
 
-1. The repository URL of each model and their specific versions (e.g., commit hash, tag, release) are specified as:
+1. The repository URL of each model and their specific versions if needed (e.g., commit hash, tag, release) are specified as:
 
     .. literalinclude:: ../../tutorials/case_h/models.yml
         :caption: tutorials/case_h/models.yml
         :language: yaml
-        :lines: 1-3, 11-13, 21-23
+        :lines: 1-3, 12-13, 22-23
 
 2. A ``path`` needs to be indicated for each model, to both download the repository contents therein and from where the source code will be executed.
 
@@ -100,7 +100,8 @@ As in :ref:`Tutorial G<case_g>`, each **Model** requires to build and execute a 
     .. literalinclude:: ../../tutorials/case_h/models.yml
         :caption: tutorials/case_h/models.yml
         :language: yaml
-        :lines: 5
+        :emphasize-lines: 6
+        :lines: 1-6
         :lineno-match:
 
     .. note::
@@ -131,7 +132,7 @@ As in :ref:`Tutorial G<case_g>`, each **Model** requires to build and execute a 
     .. literalinclude:: ../../tutorials/case_h/models.yml
         :caption: tutorials/case_h/models.yml
         :language: yaml
-        :lines: 1,6,11,15,21,25
+        :lines: 1,7,12,15,22,25
 
     .. important::
         Please refer to :ref:`Tutorial G<case_g>` for example of how to set up ``func`` for the model and interface it to **floatCSEP**.
@@ -141,13 +142,13 @@ As in :ref:`Tutorial G<case_g>`, each **Model** requires to build and execute a 
     .. literalinclude:: ../../tutorials/case_h/models.yml
         :caption: tutorials/case_h/models.yml
         :language: yaml
-        :lines: 21, 26
+        :lines: 12,16,22, 26
 
     The experiment will read the forecasts as:
 
     .. code-block::
 
-        {model_path}/{forecasts}/{prefix}_{start}_{end}.csv
+        {model_path}/forecasts/{prefix}_{start}_{end}.csv
 
     where ``start`` and ``end`` follow either the ``%Y-%m-%dT%H:%M:%S.%f`` - ISO861 FORMAT, or the short date version ``%Y-%m-%d`` if the windows are set by midnight.
 
@@ -156,8 +157,11 @@ As in :ref:`Tutorial G<case_g>`, each **Model** requires to build and execute a 
     .. literalinclude:: ../../tutorials/case_h/models.yml
         :caption: tutorials/case_h/models.yml
         :language: yaml
-        :lines: 11,17-20,21,27-31
+        :lines: 12,18-21,22,28-33
 
+.. note::
+
+    For further details on how to configure time-dependent models, see :ref:`model_config`
 
 Time
 ~~~~
@@ -172,7 +176,7 @@ Time
 Catalog
 ~~~~~~~
 
-    The catalog was obtained *prior* to the experiment using ``query_bsi``, but it was filtered from 2006 onwards, so it has enough data for the model calibration.
+    The catalog was obtained *prior* to the experiment using :func:`query bsi <csep.query_bsi>`, but it was filtered from 2006 onwards, so it has enough data for the model calibration.
 
 
 Tests
@@ -213,6 +217,11 @@ Custom Post-Process
 
[…]
