+++ tutorials/case_e/models.yml
@@ -1,12 +1,9 @@
 - ALM:
     path: models/gulia-wiemer.ALM.italy.10yr.2010-01-01.xml
     forecast_unit: 10
-    store_db: True
 - MPS04:
     path: models/meletti.MPS04.italy.10yr.2010-01-01.xml
     forecast_unit: 10
-    store_db: True
 - TripleS-CPTI:
     path: models/zechar.TripleS-CPTI.italy.10yr.2010-01-01.xml
     forecast_unit: 10
-    store_db: True
