### 2026-01-29 cseptesting/floatcsep#122 @pabloitu
Update v0.5.0
Stable release including all JOSS reviews (openjournals/joss-reviews#9408). Version 0.5.0 should be entirely compatible with ```pyCSEP``` v0.8.0 (See PR cseptesting/pycsep#282)
