### 2026-01-29 #9408 @pabloitu
Thanks for the ping and sorry for the slow reply; it took a bit to reproduce and debug the issues that Kelian reported. Here’s what we found and what we changed:

**Issues**

- There was a bug in the ```conda``` recipe. On Python 3.12, ```conda``` installs would fall back to an older floatcsep version. This led to failures when running the tutorials with an older release. Fixed in conda-forge/floatcsep-feedstock@3fc3157611b5d5b01062994026d174cdc587178d

- There were recent dependency changes, which caused tutorials to fail (notably scipy, matplotlib, and the Zenodo API), affecting both pycsep and floatcsep.
  - pycsep fixes: cseptesting/pycsep#285 and cseptesting/pycsep#290
  - floatcsep fix: cseptesting/floatcsep#118
  - A few Windows-specific issues were also resolved in the above PRs.

**Docs/review feedback:**

We strengthened the connection between floatcsep tutorials and the relevant pycsep concepts/docs in two ways:

- We added/adjusted inline references at the points where pycsep concepts are introduced.

- We added a short **pyCSEP under the hood** section under every tutorial listing the exact pycsep functions/classes/methods used.

These documentation updates are in cseptesting/floatcsep#121.

**New stable release:**

These fixes will land in a stable floatcsep release (targeting v0.5.0; see cseptesting/floatcsep#122) after pycsep v0.8.0 is released and available on conda-forge (see cseptesting/pycsep#282).

If all goes as planned with packaging, all required changes should be included early next week, once the pycsep release has propagated and the new floatcsep release is out. Or to hasten the process, the latest version of floatcsep could also be used ([Installation instructions](https://floatcsep.readthedocs.io/en/latest/intro/installation.html#latest-version)), although some pyCSEP errors would appear.

Thank you again for the reviews and bringing our attention to the running issues.
