### 2026-02-03 3990389
ft: updated interfaces to pycsep v0.8.0: modified plot function names from all tutorials, modified pycsep function parsers, and fixed imports.
