### 2026-02-03 5a8a6e9
ci: updated workflows and removed py=3.9 support fix: changed tutorial g env to conda.
