### 2026-02-03 0f0b5a9
Update floatCSEP interface to pyCSEP v0.8.0 (#123)
* ft: updated interfaces to pycsep v0.8.0: modified plot function names from all tutorials, modified pycsep function parsers, and fixed imports.

* ci: updated workflows and removed py=3.9 support
fix: changed tutorial g env to conda.

* fix: case g is now tested only locally
