+++ tutorials/case_d/tests.yml
@@ -1,6 +1,6 @@
 - Poisson S:
     func: poisson_evaluations.spatial_test
-    plot_func: plot_poisson_consistency_test
+    plot_func: plot_consistency_test
     plot_args:
         title: S-test
         xlabel: Spatial Log-Likelihood
