+++ pyproject.toml
@@ -17,7 +17,7 @@ keywords = [
 authors = [
     { name = "Pablo Iturrieta", email = "pciturri@gfz-potsdam.de" },
 ]
-requires-python = ">=3.9,<3.13"
+requires-python = ">=3.10,<3.13"
 classifiers = [
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.9",
