### 2026-02-04 #9408 @pabloitu
@andreww, @keliankaz, along with cseptesting/floatcsep#121 addressing the required documentation changes, the new `floatCSEP` *stable* release was published with cseptesting/floatcsep#124. It is already available in `conda-forge`. All aforementioned bug fixes were included, and it is ready for its tutorials to be tested. I tested them on the CI and local machines. In any case, the **latest version** could also be used to run the tutorial set (installation instructions are [here](https://floatcsep.readthedocs.io/en/latest/intro/installation.html#latest-version)).
