### 2026-02-16 #9408 @andreww
> Alright, I tested the tutorial and they work for me. As far as I am concerned this is good to go!

Excellent - thank's @keliankaz. I'll move onto the post-review tasks I need to do ahead of formally recommending acceptance. It would be great if you could tick the last two boxes in https://github.com/openjournals/joss-reviews/issues/9408#issuecomment-3697225275
