### 2026-02-16 #9408 @andreww
@pabloitu, at this point could you:
- Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct). You may also add the authors' ORCID.
