### 2026-02-16 #9408 @pabloitu
Authors and corresponding ORCID were checked! I've released a new tagged version, v0.5.1 (https://github.com/cseptesting/floatcsep/releases/tag/v0.5.1), which is linked to Zenodo automatically (All versions: https://doi.org/10.5281/zenodo.7953816 ; Latest version: https://doi.org/10.5281/zenodo.18662328). The Zenodo metadata matches the title and authors.

Also, thank you @keliankaz and @sscivier for all the constructive reviews! They significantly helped us improve the quality of our software and repository.
