### 2026-02-16 #9408 @andreww
> Authors and corresponding ORCID were checked! I've released a new tagged version, v0.5.1 (https://github.com/cseptesting/floatcsep/releases/tag/v0.5.1), which is linked to Zenodo automatically (All versions: https://doi.org/10.5281/zenodo.7953816 ; Latest version: https://doi.org/10.5281/zenodo.18662328). The Zenodo metadata matches the title and authors.
>
> Also, thank you [@keliankaz](https://github.com/keliankaz) and [@sscivier](https://github.com/sscivier) for all the constructive reviews! They significantly helped us improve the quality of our software and repository.

Super - I found some minor typographic stuff while checking the references (mostly capitalisation of countries being lost due to lack of protection in BiBtex) - some fixes in https://github.com/cseptesting/floatcsep/pull/125 - but I may have missed some. I still have to give the text of the paper a proper final read through.
