### 2026-02-17 #9408 @andreww
@pabloitu - I've read through the paper. Beyond those bits in the references (https://github.com/cseptesting/floatcsep/pull/125) the only bits I would ask you to think about are: (1) do you need the comma in the title (I'm not sure it's needed), (2) "time scales", or "timescales" on line 31 (I think this is now mostly compounded in this usage - and two words when geologists are describing tools to define time), and (3) do you want to add a reference for ObsPy on line 71 (I'm not sure it's needed)?
