### 2026-02-18 dab125b
Protect case in paper.bib (#125)
Upper case characters in article titles sometimes need wrapping in braces in BiBTeX. I think I found most of these, but do check!
