+++ paper.bib
@@ -23,7 +23,7 @@
   publisher={Oxford University Press}
 }
 @article{bayona2022prospective,
-  title={Prospective evaluation of multiplicative hybrid earthquake forecasting models in California},
+  title={Prospective evaluation of multiplicative hybrid earthquake forecasting models in {C}alifornia},
   author={Bayona, José A and Savran, William H and Rhoades, David A and Werner, Maximilian J},
   journal={Geophysical Journal International},
   volume={229},
@@ -35,7 +35,7 @@
   publisher={Oxford University Press}
 }
 @article{bayona2023regionally,
-  title={Are regionally calibrated seismicity models more informative than global models? Insights from California, New Zealand, and Italy},
+  title={Are regionally calibrated seismicity models more informative than global models? Insights from {C}alifornia, {N}ew {Z}ealand, and {I}taly},
   author={Bayona, José A and Savran, William H and Iturrieta, Pablo and Gerstenberger, Matthew C and Graham, Kenny M and Marzocchi, Warner and Schorlemmer, Danijel and Werner, Maximilian J},
   journal={The Seismic Record},
   volume={3},
@@ -47,7 +47,7 @@
   publisher={Seismological Society of America}
 }
 @article{brehmer2025enhancing,
-  title={Enhancing the Statistical Evaluation of Earthquake Forecasts—An Application to Italy},
+  title={Enhancing the Statistical Evaluation of Earthquake Forecasts—An Application to {I}taly},
   author={Brehmer, Jonas R and Kraus, Kristof and Gneiting, Tilmann and Herrmann, Marcus and Marzocchi, Warner},
   journal={Seismological Research Letters},
   volume={96},
@@ -59,7 +59,7 @@
   publisher={Seismological Society of America}
 }
 @article{eberhard2012prospective,
-  title={A prospective earthquake forecast experiment in the western Pacific},
+  title={A prospective earthquake forecast experiment in the western {P}acific},
   author={Eberhard, David AJ and Zechar, J Douglas and Wiemer, Stefan},
   journal={Geophysical Journal International},
   volume={190},
@@ -95,7 +95,7 @@
   publisher={Blackwell Publishing Ltd Oxford, UK}
 }
 @article{gerstenberger2010new,
-  title={New Zealand earthquake forecast testing centre},
+  title={{N}ew {Z}ealand earthquake forecast testing centre},
   author={Gerstenberger, Matthew C and Rhoades, David A},
   journal={Pure and Applied Geophysics},
   volume={167},
@@ -119,7 +119,7 @@
   publisher={Seismological Society of America}
 }
 @article{iturrieta2024evaluation,
-  title={Evaluation of a decade-long prospective earthquake forecasting experiment in Italy},
+  title={Evaluation of a decade-long prospective earthquake forecasting experiment in {I}taly},
   author={Iturrieta, Pablo and Bayona, José A and Werner, Maximilian J and Schorlemmer, Danijel and Taroni, Matteo and Falcone, Giuseppe and Cotton, Fabrice and Khawaja, Asim M and Savran, William H and Marzocchi, Warner},
   journal={Seismological Research Letters},
   volume={95},
@@ -168,7 +168,7 @@
 }
 @article{kagan1994long,
   title={Long-term probabilistic forecasting of earthquakes},
-  author={Kagan, YY and Jackson, DD},
+  author={Kagan, Y Y and Jackson, D D},
   journal={Journal of Geophysical Research: Solid Earth},
   volume={99},
   number={B7},
@@ -179,7 +179,7 @@
   publisher={Wiley Online Library}
 }
 @article{asim2023multi,
-  title={Multi-resolution grids in earthquake forecasting: The Quadtree approach},
[…]
