### 2026-02-18 766d131
paper: fixed capitalization on references
