+++ paper.bib
@@ -168,7 +168,7 @@
 }
 @article{kagan1994long,
   title={Long-term probabilistic forecasting of earthquakes},
-  author={Kagan, Y Y and Jackson, D D},
+  author={Kagan, Yan Y and Jackson, David D},
   journal={Journal of Geophysical Research: Solid Earth},
   volume={99},
   number={B7},
@@ -306,7 +306,7 @@ url          = {https://doi.org/10.5281/zenodo.7584575}
   publisher={Seismological Society of America}
 }
 @article{schorlemmer2018collaboratory,
-  title={The collaboratory for the study of earthquake predictability: Achievements and priorities},
+  title={The {C}ollaboratory for the {S}tudy of {E}arthquake {P}redictability: Achievements and priorities},
   author={Schorlemmer, Danijel and Werner, Maximilian J and Marzocchi, Warner and Jordan, Thomas H and Ogata, Yosihiko and Jackson, David D and Mak, Sum and Rhoades, David A and Gerstenberger, Matthew C and Hirata, Naoshi and others},
   journal={Seismological Research Letters},
   volume={89},
@@ -354,7 +354,7 @@ url          = {https://doi.org/10.5281/zenodo.7584575}
   publisher={Oxford University Press}
 }
 @article{taroni2018prospective,
-  title={Prospective CSEP evaluation of 1-day, 3-month, and 5-yr earthquake forecasts for Italy},
+  title={Prospective CSEP evaluation of 1-day, 3-month, and 5-yr earthquake forecasts for {I}taly},
   author={Taroni, Matteo and Marzocchi, Warner and Schorlemmer, Danijel and Werner, Maximilian J and Wiemer, Stefan and Zechar, Jeremy Douglas and Heiniger, Lukas and Euchner, Fabian},
   journal={Seismological Research Letters},
   volume={89},
@@ -390,7 +390,7 @@ url          = {https://doi.org/10.5281/zenodo.7584575}
   publisher={Editrice Compositori srl}
 }
 @article{zechar2010collaboratory,
-  title={The Collaboratory for the Study of Earthquake Predictability perspective on computational earthquake science},
+  title={The {C}ollaboratory for the {S}tudy of {E}arthquake {P}redictability perspective on computational earthquake science},
   author={Zechar, J Douglas and Schorlemmer, Danijel and Liukis, Maria and Yu, John and Euchner, Fabian and Maechling, Philip J and Jordan, Thomas H},
   journal={Concurrency and Computation: Practice and Experience},
   volume={22},
