### 2026-02-18 e5d366c
paper: minor fixes
