+++ paper.bib
@@ -46,6 +46,16 @@
   url={https://doi.org/10.1785/0320230006},
   publisher={Seismological Society of America}
 }
+@article{beyreuther2010obspy,
+  title={{O}bs{P}y: A {P}ython toolbox for seismology},
+  author={Beyreuther, Moritz and Barsch, Robert and Krischer, Lion and Megies, Tobias and Behr, Yannik and Wassermann, Joachim},
+  journal={Seismological Research Letters},
+  volume={81},
+  number={3},
+  pages={530--533},
+  year={2010},
+  publisher={Seismological Society of America}
+}
 @article{brehmer2025enhancing,
   title={Enhancing the Statistical Evaluation of Earthquake Forecasts—An Application to {I}taly},
   author={Brehmer, Jonas R and Kraus, Kristof and Gneiting, Tilmann and Herrmann, Marcus and Marzocchi, Warner},
