+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'floatCSEP: An application to deploy and conduct reproducible, prospective earthquake forecasting experiments'
+title: 'floatCSEP: An application to deploy and conduct reproducible prospective earthquake forecasting experiments'
 tags:
   - Python
   - seismology
@@ -47,7 +47,7 @@ authors:
     orcid: 0000-0002-2430-2631
     affiliation: 4
 affiliations:
-  - name: GFZ Helmholtz Centre for Geosciences, Potsdam, Germany,
+  - name: GFZ Helmholtz Centre for Geosciences, Potsdam, Germany
     index: 1
   - name: University of Nevada, Reno, United States
     index: 2
@@ -75,11 +75,11 @@ Ultimately, the goal of floatCSEP is to support new official CSEP experiments an
 
 # Background
 
-Earthquake forecasts are probabilistic statements about future earthquake occurrence [@jordan2011operational], used for informing building codes, emergency response planning, and risk reduction strategies.
+Earthquake forecasts are probabilistic statements about future earthquake occurrence [@jordan2011operational] and are used to inform building codes, emergency response planning, and risk reduction strategies.
 Because earthquake occurrence is driven by complex and highly non-linear processes [e.g., @geller1997earthquake; @kagan1994observational], forecasts should be expressed and evaluated in a probabilistic framework designed to describe their fundamental uncertainties [@kagan1994long].
-To assess their reliability, further challenges are the large time scales required to collect sufficient observations for evaluation (especially of large earthquakes) and the multiple subjective biases from any post-hoc adjustments in modeling or evaluation environments [e.g., @schorlemmer2007relm].
-To address such challenges, the Collaboratory for the Study of Earthquake Predictability (CSEP) was established to facilitate rigorous, prospective forecasting experiments where all forecasting models, data sources, evaluation metrics, and related software are defined prior to the evaluation time period [e.g., @jordan2006earthquake; @schorlemmer2018collaboratory].
-CSEP experiments were carried out in so-called Testing Centers, i.e., hardware and software infrastructure designed to ensure (i) controlled access, (ii) reproducible environments for automated forecast generation, and (iii) long-term archiving of input data, metadata and results [@zechar2010collaboratory].
+Assessing forecast reliability is further challenged by the large timescales required to collect sufficient observations for evaluation (especially of large earthquakes) and the multiple subjective biases from any post-hoc adjustments in modeling or evaluation environments [e.g., @schorlemmer2007relm].
+To address such challenges, the Collaboratory for the Study of Earthquake Predictability (CSEP) was established to facilitate rigorous, prospective forecasting experiments where all forecasting models, data sources, evaluation metrics, and related software are defined prior to the evaluation period [e.g., @jordan2006earthquake; @schorlemmer2018collaboratory].
+CSEP experiments were carried out in so-called Testing Centers, i.e., hardware and software infrastructure designed to ensure (i) controlled access, (ii) reproducible environments for automated forecast generation, and (iii) long-term archiving of input data, metadata, and results [@zechar2010collaboratory].
 With this framework, CSEP has successfully hosted and published experiments across diverse geographic regions, such as California, New Zealand, Italy, Japan, and globally [@bayona2021two; @bayona2022prospective; @bayona2023regionally; @eberhard2012prospective; @field2007overview; @gerstenberger2010new; @iturrieta2024evaluation; @nanjo2011overview; @strader2018prospective; @taroni2014assessing; @taroni2018prospective; @tsuruoka2012csep; @werner2010retrospective; @zechar2013regional].
 These efforts have substantially advanced scientific rigor and established community standards in earthquake forecasting research, thereby contributing to better forecasts and seismic hazard assessments [e.g., @michael2018preface; @schorlemmer2018collaboratory].
 
[…]
