### 2026-02-18 #9408 @pabloitu
@andreww super.

I removed the comma for the title: the comma was used to reduce the "and"s in the title, but I believe it now works with stating "reproducible prospective earthquake forecasting experiments" (changes in cseptesting/floatcsep@e5d366c08bcbdcd65053ea62b28c74544ede319e , cseptesting/floatcsep@5488092d4e51d9abaae3837b2a8cb4899d5662a8, and in Zenodo).

You are right, timescales are mostly used in geology/seismology contexts. Modified in cseptesting/floatcsep@e5d366c08bcbdcd65053ea62b28c74544ede319e. Added obspy reference in the same commit.

Note: unit tests are failing due to an obspy issue. We are considering removing obspy from pyCSEP dependencies, and reimplement the limited usage we have from that library, but for now we prefer to wait they fix the issue (https://github.com/obspy/obspy/issues/3681).
