### 2026-02-18 #9408 @andreww
@pabloitu  - all looks good to me except the editoralbot has spotted that you (and I) missed a DOI on the new reference for ObsPy. (The suggested DOI: 10.1785/gssrl.81.3.530 is right - if you could add that then I think an editor in chief will be able to regenerate the PDF)
