### 2026-02-19 0b9bd12
paper: added doi for ObsPy reference.
