+++ paper.bib
@@ -54,6 +54,7 @@
   number={3},
   pages={530--533},
   year={2010},
+  doi = {10.1785/gssrl.81.3.530},
   publisher={Seismological Society of America}
 }
 @article{brehmer2025enhancing,
