### 2026-02-19 #9408 @pabloitu
@andreww Oh I'm sorry I didn't checked that. I added the corresponding DOI in commit cseptesting/floatcsep@0b9bd122582803e6b6833f0ee1c1ea47057759f3 . Hope that to regenerate the final PDF is not so much of a problem
