### 2026-02-20 #9408 @andreww
> [@andreww](https://github.com/andreww) Oh I'm sorry I didn't checked that. I added the corresponding DOI in commit [cseptesting/floatcsep@0b9bd12](https://github.com/cseptesting/floatcsep/commit/0b9bd122582803e6b6833f0ee1c1ea47057759f3) . Hope that to regenerate the final PDF is not so much of a problem

Thanks - all looks good now.
